models were refactored #11
