+++ organizations/models/__init__.py
@@ -1,0 +1,16 @@
+from .address import Address
+from .contact import Contact
+from .industry import Industry
+from .industry_resource import industries_resources_table
+from .message import Message
+from .methodicaldoc import MethodicalDoc
+from .methodicaldoc_message import methodicaldocs_messages
+from .okrug import Okrug
+from .orgadmdoc import OrgAdmDoc
+from .orgadmdoc_organization import OrgAdmDocOrganization
+from .organization import Organization
+from .organization_message import organizations_messages
+from .region import Region
+from .regions_resource import regions_resources_table
+from .resource import Resource
+from .responsibility import Responsibility
