+++ organizations/models/industry_resource.py
@@ -1,0 +1,15 @@
+from ..extentions import db
+
+industries_resources_table = db.Table(
+    "industries_resources",
+    db.Column("industry_id",
+              db.Integer,
+              db.ForeignKey("industries.industry_id",
+                            ondelete="CASCADE")
+              ),
+    db.Column("resource_id",
+              db.Integer,
+              db.ForeignKey("resources.resource_id",
+                            ondelete="CASCADE")
+              )
+)
